Present all functionality as middleware, ServerlessMiddleware base, optionally handle debug responses in the middleware (if implemented for platform), update readmes and guides
