+++ mangum/__version__.py
@@ -1,1 +1,1 @@
-__version__ = "0.5.3"  # pragma: no cover
+__version__ = "0.5.4"  # pragma: no cover
