+++ tests/test_aws_middleware.py
@@ -2,7 +2,6 @@ import base64
 from starlette.applications import Starlette
 from starlette.requests import Request
 from starlette.responses import PlainTextResponse
-from mangum.platforms.aws.adapter import run_asgi
 from mangum.platforms.aws.middleware import AWSLambdaMiddleware
 
 
@@ -15,7 +14,8 @@ def test_aws_response(mock_data) -> None:
         return asgi
 
     mock_event = mock_data.get_aws_event()
-    response = run_asgi(app, mock_event, {})
+    handler = AWSLambdaMiddleware(app)
+    response = handler(mock_event, {})
 
     assert response == {
         "statusCode": 200,
@@ -39,16 +39,9 @@ def test_aws_response_with_body(mock_data) -> None:
         return asgi
 
     mock_event = mock_data.get_aws_event()
-    response = run_asgi(app, mock_event, {})
+    handler = AWSLambdaMiddleware(app)
+    response = handler(mock_event, {})
 
-    assert response == {
-        "statusCode": 200,
-        "isBase64Encoded": False,
-        "headers": {"content-length": "3", "content-type": "text/plain; charset=utf-8"},
-        "body": "123",
-    }
-
-    response = AWSLambdaMiddleware(app)(mock_event, {})
     assert response == {
         "statusCode": 200,
         "isBase64Encoded": False,
@@ -72,7 +65,8 @@ def test_aws_binary_response_with_body(mock_data) -> None:
     body_encoded = base64.b64encode(body)
     mock_event["body"] = body_encoded
     mock_event["isBase64Encoded"] = True
-    response = run_asgi(app, mock_event, {})
+    handler = AWSLambdaMiddleware(app)
+    response = handler(mock_event, {})
 
     assert response == {
         "statusCode": 200,
@@ -92,7 +86,8 @@ def test_starlette_aws_response(mock_data) -> None:
     def homepage(request):
         return PlainTextResponse("Hello, world!")
 
-    response = run_asgi(app, mock_event, {})
+    handler = AWSLambdaMiddleware(app)
+    response = handler(mock_event, {})
 
     assert response == {
         "statusCode": 200,
